\name{fa.data}
\docType{data}
\alias{fa.data}
\alias{fa.data2}
\title{
  Data sets
}

\description{
  Example fa dataset obtained by gas chromatography with flame
  ionization detection.
}

\usage{
fa.data

}

\format{
  a data.frame
}


\examples{
fa.data
plotchrom(fa.data)
}
\references{
}

\keyword{datasets}
